Moved contact routes into contact folder. Migrated SafeRouter to TS and added type checking to it.
