+++ hrm-service/src/contact/contact-routes-v0.ts
@@ -1,9 +1,9 @@
-const models = require('../../models');
-const { SafeRouter, publicEndpoint } = require('../../permissions');
-const createError = require('http-errors');
-
+import models from '../models';
+import { SafeRouter, publicEndpoint } from '../permissions';
+import createError from 'http-errors';
+import contactControllerFactory from '../controllers/contact-controller';
 const { Contact } = models;
-const ContactController = require('../../controllers/contact-controller')(Contact);
+const ContactController = contactControllerFactory(Contact);
 
 const contactsRouter = SafeRouter();
 
@@ -46,4 +46,11 @@ contactsRouter.post('/search', publicEndpoint, async (req, res) => {
   res.json(searchResults);
 });
 
-module.exports = contactsRouter.expressRouter;
+contactsRouter.patch('/:contactId', publicEndpoint, async (req, res) => {
+  const { accountSid, user } = req;
+
+  const contact = await ContactController.createContact(req.body, accountSid, user.workerSid);
+  res.json(contact);
+})
+
+export default contactsRouter.expressRouter;
