+++ hrm-service/src/permissions/index.ts
@@ -1,6 +1,4 @@
-const User = require('./user');
-const { SafeRouter, publicEndpoint } = require('./safe-router');
-const { setupCanForRules } = require('./setupCanForRules');
+import { setupCanForRules } from './setupCanForRules';
 
 const openRules = require('../../permission-rules/open.json');
 const brRules = require('../../permission-rules/br.json');
@@ -12,8 +10,6 @@ const mwRules = require('../../permission-rules/mw.json');
 const ukRules = require('../../permission-rules/uk.json');
 const zaRules = require('../../permission-rules/za.json');
 const zmRules = require('../../permission-rules/zm.json');
-
-const { canEditCase, canViewPostSurvey } = require('./middlewares');
 
 // TODO: maybe factor out so it's easier to build the API that retrieves this to the frontend
 const rulesMap = {
@@ -49,7 +45,7 @@ const initializedCanOpenRules = setupCanForRules(openRules);
  * @param {string} permissionsConfig
  * @throws Will throw if initializedCan is not a function
  */
-const applyPermissions = (req, initializedCan, permissionsConfig) => {
+export const applyPermissions = (req, initializedCan, permissionsConfig) => {
   if (!initializedCan) throw new Error(`Cannot find rules for ${permissionsConfig}`);
 
   if (typeof initializedCan === 'string')
@@ -61,7 +57,7 @@ const applyPermissions = (req, initializedCan, permissionsConfig) => {
   req.can = initializedCan;
 };
 
-const setupPermissions = (req, res, next) => {
+export const setupPermissions = (req, res, next) => {
   if (process.env.USE_OPEN_PERMISSIONS) {
     applyPermissions(req, initializedCanOpenRules, 'open rules');
     return next();
@@ -76,11 +72,6 @@ const setupPermissions = (req, res, next) => {
   return next();
 };
 
-module.exports = {
-  setupPermissions,
-  User,
-  SafeRouter,
-  publicEndpoint,
-  canEditCase,
-  canViewPostSurvey,
-};
+export * from './middlewares';
+export * from './safe-router';
+export * from './user';
