+++ hrm-service/src/permissions/user.js
@@ -1,8 +1,6 @@
-const User = class {
+export const User = class {
   constructor(workerSid, roles) {
     this.workerSid = workerSid;
     this.roles = roles;
   }
 };
-
-module.exports = User;
