+++ hrm-service/src/routes/v0/index.ts
@@ -1,6 +1,6 @@
 import cases from '../../case/case-routes-v0';
+import contacts from '../../contact/contact-routes-v0';
 const { Router } = require('express');
-const contacts = require('./contacts');
 const postSurveys = require('./post-surveys');
 const csamReports = require('./csam-reports');
